+++ src/colors.rs
@@ -1,16 +1,16 @@
 
 use crate::vectors::Color;
 
-pub fn clamp(x: f64, min: f64, max: f64) -> f64 {
+pub fn clamp(x: f32, min: f32, max: f32) -> f32 {
         if x < min {min} else if x > max {max} else {x}
 }
 
 pub fn write_color(pixels: Color, samples_per_pixel: u32) {
-        let scale: f64 = 1.0 / (samples_per_pixel as f64);
+        let scale: f32 = 1.0 / (samples_per_pixel as f32);
 
-        let r: f64 = (pixels.x * scale).sqrt();
-        let b: f64 = (pixels.y * scale).sqrt();
-        let g: f64 = (pixels.z * scale).sqrt();
+        let r: f32 = (pixels.x * scale).sqrt();
+        let b: f32 = (pixels.y * scale).sqrt();
+        let g: f32 = (pixels.z * scale).sqrt();
 
 
         println!("{} {} {}",
